+++ packages/common/constants/index.js
@@ -15,8 +15,8 @@ const DragItemTypes = {
 };
 
 const CellExpand = {
-  DOWN_TRIANGLE: String.fromCharCode('9660'),
-  RIGHT_TRIANGLE: String.fromCharCode('9654')
+  DOWN_TRIANGLE: String.fromCharCode(9660),
+  RIGHT_TRIANGLE: String.fromCharCode(9654)
 };
