+++ packages/common/utils/index.js
@@ -1,25 +1,50 @@
-import { List } from 'immutable';
+import { List, Iterable, Map } from 'immutable';
 
-module.exports = {
-  isEmptyArray: require('./isEmptyArray'),
-  isEmptyObject: require('./isEmptyObject'),
-  isImmutableCollection: require('./isImmutableCollection'),
-  getMixedTypeValueRetriever: require('./mixedTypeValueRetriever'),
-  isColumnsImmutable: require('./isColumnsImmutable'),
-  isImmutableMap: require('./isImmutableMap'),
-  last: (arrayOrList) => {
-    if (arrayOrList == null) {
-      throw new Error('arrayOrCollection is null');
-    }
+export const isColumnsImmutable = (columns) => {
+  return (typeof Immutable !== 'undefined' && (columns instanceof Immutable.List));
+};
 
-    if (List.isList(arrayOrList)) {
-      return arrayOrList.last();
-    }
+export const isEmptyArray = (obj) => {
+  return Array.isArray(obj) && obj.length === 0;
+};
 
-    if (Array.isArray(arrayOrList)) {
-      return arrayOrList[arrayOrList.length - 1];
-    }
+export const isFunction = (functionToCheck) => {
+  let getType = {};
+  return functionToCheck && getType.toString.call(functionToCheck) === '[object Function]';
+};
 
-    throw new Error('Cant get last of: ' + typeof(arrayOrList));
+export const isEmptyObject = (obj) => {
+  return Object.keys(obj).length === 0 && obj.constructor === Object;
+};
+
+export const isImmutableCollection = objToVerify => {
+  return Iterable.isIterable(objToVerify);
+};
+
+export const getMixedTypeValueRetriever = (isImmutable) => {
+  let retObj = {};
+  const retriever = (item, key) => { return item[key]; };
+  const immutableRetriever =  (immutable, key) => { return immutable.get(key); };
+
+  retObj.getValue = isImmutable ? immutableRetriever : retriever;
+
+  return retObj;
+};
+
+export const isImmutableMap = Map.isMap;
+
+export const last = arrayOrList => {
+  if (arrayOrList == null) {
+    throw new Error('arrayOrCollection is null');
   }
+
+  if (List.isList(arrayOrList)) {
+    return arrayOrList.last();
+  }
+
+  if (Array.isArray(arrayOrList)) {
+    return arrayOrList[arrayOrList.length - 1];
+  }
+
+  throw new Error('Cant get last of: ' + typeof(arrayOrList));
 };
