--- packages/common/utils/isColumnsImmutable.js
@@ -1,3 +1,0 @@
-export default function isColumnsImmutable(columns) {
-  return (typeof Immutable !== 'undefined' && (columns instanceof Immutable.List));
-}
