--- packages/common/utils/isEmptyArray.js
@@ -1,5 +1,0 @@
-const isEmptyArray = (obj) => {
-  return Array.isArray(obj) && obj.length === 0;
-};
-
-export default isEmptyArray;
