--- packages/common/utils/isEmptyObject.js
@@ -1,5 +1,0 @@
-export default function isEmpty(obj) {
-  return Object.keys(obj).length === 0 && obj.constructor === Object;
-}
-
-
