--- packages/common/utils/isFunction.js
@@ -1,4 +1,0 @@
-export default function isFunction(functionToCheck) {
-  let getType = {};
-  return functionToCheck && getType.toString.call(functionToCheck) === '[object Function]';
-}
