--- packages/common/utils/isImmutableCollection.js
@@ -1,7 +1,0 @@
-import { Iterable } from 'immutable';
-
-export default function isImmutableCollection(objToVerify) {
-  return Iterable.isIterable(objToVerify);
-};
-
-
