--- packages/common/utils/isImmutableMap.js
@@ -1,3 +1,0 @@
-import { Map } from 'immutable';
-
-module.exports = Map.isMap;
