--- packages/common/utils/mixedTypeValueRetriever.js
@@ -1,9 +1,0 @@
-export default function getMixedTypeValueRetriever(isImmutable) {
-  let retObj = {};
-  const retriever = (item, key) => { return item[key]; };
-  const immutableRetriever =  (immutable, key) => { return immutable.get(key); };
-
-  retObj.getValue = isImmutable ? immutableRetriever : retriever;
-
-  return retObj;
-}
