+++ packages/react-data-grid-addons/src/cells/headerCells/filters/AutoCompleteFilter.js
@@ -2,7 +2,7 @@ import 'react-select/dist/react-select.css';
 import React from 'react';
 import PropTypes from 'prop-types';
 import Select from 'react-select';
-import isEmptyArray from 'common/utils/isEmptyArray';
+import {isEmptyArray} from 'common/utils';
 import ExcelColumn from 'common/prop-shapes/ExcelColumn';
 
 class AutoCompleteFilter extends React.Component {
