+++ packages/react-data-grid-addons/src/data/RowGrouper.js
@@ -1,4 +1,4 @@
-import isImmutableCollection from 'common/utils/isImmutableCollection';
+import {isImmutableCollection} from 'common/utils';
 import Resolver from './RowGrouperResolver';
