+++ packages/react-data-grid-addons/src/draggable/DragDropContainer.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import html5DragDropContext from '../shared/html5DragDropContext';
 import DraggableHeaderCell from './DraggableHeaderCell';
 import RowDragLayer from './RowDragLayer';
-import isColumnsImmutable from 'common/utils/isColumnsImmutable';
+import {isColumnsImmutable} from 'common/utils';
 import PropTypes from 'prop-types';
 
 class DraggableContainer extends Component {
