+++ packages/react-data-grid/src/Cell.js
@@ -2,12 +2,12 @@ const React = require('react');
 import PropTypes from 'prop-types';
 const joinClasses = require('classnames');
 import ExcelColumn from 'common/prop-shapes/ExcelColumn';
-import isFunction from 'common/utils/isFunction';
+import {isFunction} from 'common/utils';
 import CellMetaDataShape from 'common/prop-shapes/CellMetaDataShape';
 const SimpleCellFormatter = require('./formatters/SimpleCellFormatter');
 const createObjectWithProperties = require('./createObjectWithProperties');
 import CellAction from './CellAction';
-import CellExpand from './CellExpand';
+import CellExpand from './CellExpander';
 import ChildRowDeleteButton from './ChildRowDeleteButton';
 import columnUtils from './ColumnUtils';
 require('../../../themes/react-data-grid-cell.css');
