+++ packages/react-data-grid/src/utils/SelectedCellUtils.js
@@ -1,5 +1,5 @@
 import { CellNavigationMode } from 'common/constants';
-import isFunction from 'common/utils/isFunction';
+import {isFunction} from 'common/utils';
 import * as rowUtils from '../RowUtils';
 import * as columnUtils from '../ColumnUtils';
 import zIndexes from 'common/constants/zIndexes';
